@@ -3,6 +3,8 @@ import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart
 import 'package:provider/provider.dart';
 
 extension PopupBuildContext on BuildContext {
+  bool get isSelectable => _watch.value.isSelectable;
+
   GlobalKey<HtmlWidgetState> get key => _watch.value.key;
 
   bool get showPerformanceOverlay => _watch.value.showPerformanceOverlay;
@@ -15,11 +17,13 @@ extension PopupBuildContext on BuildContext {
 class PopupMenu extends StatelessWidget {
   final bool scrollToTop;
   final bool showPerfOverlay;
+  final bool toggleIsSelectable;
 
   const PopupMenu({
     Key key,
     this.scrollToTop = false,
     this.showPerfOverlay = true,
+    this.toggleIsSelectable = false,
   }) : super(key: key);
 
   @override
@@ -37,6 +41,9 @@ class PopupMenu extends StatelessWidget {
             notifier.value = value.copyWith(
               showPerformanceOverlay: !value.showPerformanceOverlay,
             );
+            break;
+          case _PopupMenuValue.toggleIsSelectable:
+            notifier.value = value.copyWith(isSelectable: !value.isSelectable);
             break;
         }
       },
@@ -58,6 +65,14 @@ class PopupMenu extends StatelessWidget {
                 value: value.showPerformanceOverlay,
               ),
             ),
+          if (toggleIsSelectable)
+            PopupMenuItem(
+              value: _PopupMenuValue.toggleIsSelectable,
+              child: _CheckBoxMenuItem(
+                title: 'isSelectable',
+                value: value.isSelectable,
+              ),
+            ),
         ];
       },
     );
@@ -66,14 +81,22 @@ class PopupMenu extends StatelessWidget {
 
 class PopupMenuStateProvider extends StatelessWidget {
   final WidgetBuilder builder;
+  final bool initialIsSelectable;
 
-  const PopupMenuStateProvider({@required this.builder, Key key})
-      : super(key: key);
+  const PopupMenuStateProvider({
+    @required this.builder,
+    this.initialIsSelectable = false,
+    Key key,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider(
-      create: (_) => ValueNotifier(_PopupMenuState()),
+      create: (_) => ValueNotifier(
+        _PopupMenuState(
+          isSelectable: initialIsSelectable,
+        ),
+      ),
       child: Builder(builder: builder),
     );
   }
@@ -103,10 +126,12 @@ class _CheckBoxMenuItem extends StatelessWidget {
 
 @immutable
 class _PopupMenuState {
+  final bool isSelectable;
   final GlobalKey<HtmlWidgetState> key;
   final bool showPerformanceOverlay;
 
   _PopupMenuState({
+    this.isSelectable = false,
     GlobalKey<HtmlWidgetState> key,
     this.showPerformanceOverlay = false,
   }) : key = key ?? GlobalKey<HtmlWidgetState>();
@@ -116,6 +141,7 @@ class _PopupMenuState {
     bool showPerformanceOverlay,
   }) =>
       _PopupMenuState(
+        isSelectable: isSelectable ?? this.isSelectable,
         showPerformanceOverlay:
             showPerformanceOverlay ?? this.showPerformanceOverlay,
       );
@@ -124,4 +150,5 @@ class _PopupMenuState {
 enum _PopupMenuValue {
   scrollToTop,
   showPerformanceOverlay,
+  toggleIsSelectable,
 }
