@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:demo_app/widgets/popup_menu.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_widget_from_html/flutter_widget_from_html.dart'
@@ -18,7 +19,9 @@ class Golden extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final baseUrl = Uri.parse('https://www.w3schools.com/html/');
-    final withEnhanced = RegExp(r'^(AUDIO|IFRAME|SVG|VIDEO)$').hasMatch(name);
+    final isSelectable = context.isSelectable;
+    final withEnhanced =
+        !isSelectable && RegExp(r'^(AUDIO|IFRAME|SVG|VIDEO)$').hasMatch(name);
 
     final children = <Widget>[
       if (withEnhanced)
@@ -36,10 +39,16 @@ class Golden extends StatelessWidget {
         ),
       LimitedBox(
         maxHeight: 400,
-        child: core.HtmlWidget(
-          html,
-          baseUrl: baseUrl,
-        ),
+        child: isSelectable
+            ? enhanced.HtmlWidget(
+                html,
+                baseUrl: baseUrl,
+                isSelectable: true,
+              )
+            : core.HtmlWidget(
+                html,
+                baseUrl: baseUrl,
+              ),
       ),
     ];
 
@@ -63,7 +72,14 @@ class Golden extends StatelessWidget {
     }
 
     return Scaffold(
-      appBar: AppBar(title: Text(name)),
+      appBar: AppBar(
+        title: Text(name),
+        actions: const [
+          PopupMenu(
+            toggleIsSelectable: true,
+          ),
+        ],
+      ),
       body: SingleChildScrollView(
         child: RepaintBoundary(
           key: targetKey,
@@ -88,7 +104,7 @@ class GoldensScreen extends StatefulWidget {
   const GoldensScreen({Key key}) : super(key: key);
 
   @override
-  _GoldensState createState() => _GoldensState();
+  State<GoldensScreen> createState() => _GoldensState();
 }
 
 class _GoldensState extends State<GoldensScreen> {
