@@ -16,7 +16,7 @@ import 'internal/platform_specific/fallback.dart'
 
 /// A factory to build widgets.
 class WidgetFactory {
-  /// Setting this property to true replaces the default loading widget with a static [Text].
+  /// Setting this property to true replaces the default with a static [Text].
   /// This property is most useful for testing purposes.
   ///
   /// Defaults to `false`, resulting in a [CircularProgressIndicator].
@@ -49,7 +49,7 @@ class WidgetFactory {
 
   /// Gets the current anchor registry.
   ///
-  /// This is an implementation detail and may be changed without a major version bump.
+  /// This is an impl detail and may be changed without a major version bump.
   @protected
   AnchorRegistry get anchorRegistry => _anchorRegistry;
 
@@ -165,6 +165,7 @@ class WidgetFactory {
 
     final container = child is Container ? child : null;
     final decoratedBox = child is DecoratedBox ? child : null;
+    final grandChild = container?.child ?? decoratedBox?.child;
     final prevDeco = container?.decoration ?? decoratedBox?.decoration;
     final baseDeco =
         prevDeco is BoxDecoration ? prevDeco : const BoxDecoration();
@@ -177,7 +178,7 @@ class WidgetFactory {
     if (!isBorderBox || container != null) {
       return Container(
         decoration: decoration,
-        child: container?.child ?? child,
+        child: grandChild ?? child,
       );
     } else {
       return DecoratedBox(
@@ -636,16 +637,18 @@ class WidgetFactory {
 
     switch (meta.element.localName) {
       case kTagA:
-        _tagA ??= TagA(this).buildOp;
-        meta.register(_tagA!);
+        if (attrs.containsKey(kAttributeAHref)) {
+          final tagA = _tagA ??= TagA(this).buildOp;
+          meta.register(tagA);
 
-        meta.tsb.enqueue(
-          _tagAColor ??= (tsh, _) => tsh.copyWith(
+          final tagAColor = _tagAColor ??= (tsh, _) => tsh.copyWith(
                 style: tsh.style.copyWith(
                   color: tsh.getDependency<ThemeData>().colorScheme.primary,
                 ),
-              ),
-        );
+              );
+
+          meta.tsb.enqueue(tagAColor);
+        }
 
         final name = attrs[kAttributeAName];
         if (name != null) {
