@@ -21,6 +21,11 @@ class HtmlTable extends MultiChildRenderObjectWidget {
   /// The companion data for table.
   final HtmlTableCompanion companion;
 
+  /// Determines the order to lay children out horizontally.
+  ///
+  /// Default: [TextDirection.ltr].
+  final TextDirection textDirection;
+
   /// Creates a TABLE widget.
   HtmlTable({
     this.border,
@@ -28,6 +33,7 @@ class HtmlTable extends MultiChildRenderObjectWidget {
     this.borderSpacing = 0.0,
     required List<Widget> children,
     required this.companion,
+    this.textDirection = TextDirection.ltr,
     Key? key,
   }) : super(children: children, key: key);
 
@@ -36,6 +42,7 @@ class HtmlTable extends MultiChildRenderObjectWidget {
         border,
         borderSpacing,
         companion,
+        textDirection,
         borderCollapse: borderCollapse,
       );
 
@@ -53,15 +60,23 @@ class HtmlTable extends MultiChildRenderObjectWidget {
     );
     properties
         .add(DoubleProperty('borderSpacing', borderSpacing, defaultValue: 0.0));
+    properties.add(
+      DiagnosticsProperty(
+        'textDirection',
+        textDirection,
+        defaultValue: TextDirection.ltr,
+      ),
+    );
   }
 
   @override
-  void updateRenderObject(BuildContext _, _TableRenderObject renderObject) {
-    renderObject
+  void updateRenderObject(BuildContext _, RenderObject renderObject) {
+    (renderObject as _TableRenderObject)
       ..border = border
       ..borderCollapse = borderCollapse
       ..borderSpacing = borderSpacing
-      ..companion = companion;
+      ..companion = companion
+      ..textDirection = textDirection;
   }
 }
 
@@ -168,13 +183,10 @@ class HtmlTableValignBaseline extends SingleChildRenderObjectWidget {
   }
 
   @override
-  void updateRenderObject(
-    BuildContext context,
-    _ValignBaselineRenderObject renderObject,
-  ) {
+  void updateRenderObject(BuildContext context, RenderObject renderObject) {
     final table = context.findAncestorWidgetOfExactType<HtmlTable>()!;
     final cell = context.findAncestorWidgetOfExactType<HtmlTableCell>()!;
-    renderObject
+    (renderObject as _ValignBaselineRenderObject)
       ..companion = table.companion
       ..row = cell.rowStart;
   }
@@ -223,7 +235,8 @@ class _TableRenderObject extends RenderBox
   _TableRenderObject(
     this._border,
     this._borderSpacing,
-    this._companion, {
+    this._companion,
+    this._textDirection, {
     required bool borderCollapse,
   }) : _borderCollapse = borderCollapse;
 
@@ -268,6 +281,17 @@ class _TableRenderObject extends RenderBox
     }
 
     _companion = v;
+    markNeedsLayout();
+  }
+
+  TextDirection _textDirection;
+  // ignore: avoid_setters_without_getters
+  set textDirection(TextDirection v) {
+    if (v == _textDirection) {
+      return;
+    }
+
+    _textDirection = v;
     markNeedsLayout();
   }
 
@@ -470,10 +494,21 @@ class _TableRenderObject extends RenderBox
       }
 
       if (child.hasSize) {
-        data.offset = Offset(
-          data.calculateX(tro, columnWidths),
-          data.calculateY(tro, rowHeights),
-        );
+        final calculatedX = data.calculateX(tro, columnWidths);
+
+        // TODO: use `late final` when https://github.com/dart-lang/coverage/issues/341 is fixed
+        late double x;
+        switch (tro._textDirection) {
+          case TextDirection.ltr:
+            x = calculatedX;
+            break;
+          case TextDirection.rtl:
+            x = calculatedWidth - childWidth - calculatedX;
+            break;
+        }
+
+        final y = data.calculateY(tro, rowHeights);
+        data.offset = Offset(x, y);
       }
     }
 
