@@ -58,7 +58,8 @@ Future<void> main() async {
       str,
       equals(
         '[Column:children='
-        '[CssBlock:child=[RichText:(:(+l+o+u:All decorations... )(:and none))]],'
+        '[CssBlock:child=[RichText:(:(+l+o+u:All decorations... )'
+        '(:and none))]],'
         '[CssBlock:child=[RichText:(:I​Like​Playing​football​​game)]],'
         '[CssBlock:child=[RichText:(:\u00A0)]]'
         ']',
@@ -1000,8 +1001,9 @@ Future<void> main() async {
             equals(
               '[Column:children='
               '[RichText:(:Foo)],'
-              '[CssSizing:$imgSizingConstraints,child=[Image:image=NetworkImage("$src", scale: 1.0)]]'
-              ']',
+              '[CssSizing:$imgSizingConstraints,child='
+              '[Image:image=NetworkImage("$src", scale: 1.0)]'
+              ']]',
             ),
           );
         }),
@@ -1414,8 +1416,9 @@ Future<void> main() async {
       expect(
         explained,
         equals(
-          '[RichText:(:(+b:bold)(: )(+w0:one)(: )(+w1:two)(: )(+w2:three)(: )(:four)(: )'
-          '(+w4:five)(: )(+w5:six)(: )(+b:seven)(: )(+w7:eight)(: )(+w8:nine))]',
+          '[RichText:(:(+b:bold)(: )(+w0:one)(: )(+w1:two)(: )(+w2:three)(: )'
+          '(:four)(: )(+w4:five)(: )(+w5:six)(: )'
+          '(+b:seven)(: )(+w7:eight)(: )(+w8:nine))]',
         ),
       );
     });
