@@ -1,10 +1,12 @@
+import 'dart:async';
+
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
 import 'package:fwfh_chewie/fwfh_chewie.dart';
-import 'package:video_player_platform_interface/messages.dart';
-import 'package:video_player_platform_interface/test.dart';
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+import 'package:video_player_platform_interface/video_player_platform_interface.dart';
 
 import '../../core/test/_.dart' as helper;
 
@@ -12,13 +14,16 @@ const kDataUri = helper.kDataUri;
 
 String? videoPlayerExplainer(helper.Explainer parent, Widget widget) {
   if (widget is VideoPlayer) {
+    final poster = widget.poster != null
+        ? ',poster=${parent.explain(widget.poster!)}'
+        : '';
     return '[VideoPlayer:url=${widget.url}'
         ',aspectRatio=${widget.aspectRatio.toStringAsFixed(2)}'
         "${!widget.autoResize ? ',autoResize=${widget.autoResize}' : ''}"
         "${widget.autoplay ? ',autoplay=${widget.autoplay}' : ''}"
         "${widget.controls ? ',controls=${widget.controls}' : ''}"
         "${widget.loop ? ',loop=${widget.loop}' : ''}"
-        "${widget.poster != null ? ',poster=${parent.explain(widget.poster!)}' : ''}"
+        '$poster'
         ']';
   }
 
@@ -58,83 +63,65 @@ Future<String> explain(
 
 void mockVideoPlayerPlatform() => _FakeVideoPlayerPlatform();
 
-class _FakeVideoPlayerPlatform extends TestHostVideoPlayerApi {
+class _FakeVideoPlayerPlatform extends Fake
+    with MockPlatformInterfaceMixin
+    implements VideoPlayerPlatform {
   var _nextTextureId = 0;
 
+  final uris = <int, String?>{};
+
   _FakeVideoPlayerPlatform() {
-    TestHostVideoPlayerApi.setup(this);
+    VideoPlayerPlatform.instance = this;
   }
 
   @override
-  TextureMessage create(CreateMessage arg) {
-    _FakeVideoEventStream(_nextTextureId, arg.uri);
-    return TextureMessage()..textureId = _nextTextureId++;
+  Widget buildView(int textureId) => const SizedBox.shrink();
+
+  @override
+  Future<int?> create(DataSource dataSource) async {
+    final textureId = _nextTextureId++;
+    uris[textureId] = dataSource.uri;
+    return textureId;
   }
 
   @override
-  void initialize() {}
+  Future<void> dispose(int textureId) async {/* intentionally left empty */}
 
   @override
-  void dispose(TextureMessage arg) {}
+  Future<void> init() async {/* intentionally left empty */}
 
   @override
-  void pause(TextureMessage arg) {}
+  Future<void> play(int textureId) async {/* intentionally left empty */}
 
   @override
-  void play(TextureMessage arg) {}
+  Future<void> pause(int textureId) async {/* intentionally left empty */}
 
   @override
-  PositionMessage position(TextureMessage arg) => PositionMessage();
-
-  @override
-  void seekTo(PositionMessage arg) {}
-
-  @override
-  void setLooping(LoopingMessage arg) {}
-
-  @override
-  void setMixWithOthers(MixWithOthersMessage arg) {}
-
-  @override
-  void setPlaybackSpeed(PlaybackSpeedMessage arg) {}
-
-  @override
-  void setVolume(VolumeMessage arg) {}
-}
-
-class _FakeVideoEventStream {
-  final int textureId;
-  final String? uri;
-
-  _FakeVideoEventStream(this.textureId, this.uri) {
-    MethodChannel(name).setMockMethodCallHandler(_handler);
+  Future<void> setLooping(int textureId, bool looping) async {
+    /* intentionally left empty */
   }
 
-  String get name => 'flutter.io/videoPlayer/videoEvents$textureId';
+  @override
+  Future<void> setPlaybackSpeed(int textureId, double speed) async {
+    /* intentionally left empty */
+  }
 
-  Future<dynamic> _handler(MethodCall call) async {
-    switch (call.method) {
-      case 'listen':
-        late ByteData byteData;
+  @override
+  Future<void> setVolume(int textureId, double volume) async {
+    /* intentionally left empty */
+  }
 
-        if (uri?.endsWith('init/error.mp4') == true) {
-          byteData = const StandardMethodCodec().encodeErrorEnvelope(
-            code: 'init_error',
-            message: uri,
-          );
-        } else {
-          final data = <String, dynamic>{
-            'event': 'initialized',
-            'duration': 1000,
-            'width': 1600,
-            'height': 900,
-          };
-          byteData = const StandardMethodCodec().encodeSuccessEnvelope(data);
-        }
-
-        final messenger = ServicesBinding.instance!.defaultBinaryMessenger;
-        await messenger.handlePlatformMessage(name, byteData, (_) {});
-        break;
+  @override
+  Stream<VideoEvent> videoEventsFor(int textureId) async* {
+    final uri = uris[textureId] ?? '';
+    if (uri.endsWith('init/error.mp4') == true) {
+      throw PlatformException(code: 'init_error', message: uri);
+    } else {
+      yield VideoEvent(
+        eventType: VideoEventType.initialized,
+        duration: const Duration(seconds: 1),
+        size: const Size(1600, 900),
+      );
     }
   }
 }
