@@ -7,9 +7,7 @@ void main() {
   const src = 'http://domain.com/video.mp4';
   const defaultAspectRatio = '1.78';
 
-  setUp(() {
-    mockVideoPlayerPlatform();
-  });
+  mockVideoPlayerPlatform();
 
   testWidgets('renders video player', (tester) async {
     const html = '<video><source src="$src"></video>';
