@@ -7,15 +7,35 @@ import '../../core/test/_.dart' as helper;
 
 String? webViewExplainer(helper.Explainer parent, Widget widget) {
   if (widget is WebView) {
+    final debuggingEnabled = widget.debuggingEnabled
+        ? ',debuggingEnabled=${widget.debuggingEnabled}'
+        : '';
+    final mediaPlaybackAlwaysAllow = widget.mediaPlaybackAlwaysAllow
+        ? ',mediaPlaybackAlwaysAllow=${widget.mediaPlaybackAlwaysAllow}'
+        : '';
+    final unsupportedWorkaroundForIssue37 =
+        !widget.unsupportedWorkaroundForIssue37
+            ? ',unsupportedWorkaroundForIssue37='
+                '${widget.unsupportedWorkaroundForIssue37}'
+            : '';
+    final unsupportedWorkaroundForIssue375 =
+        !widget.unsupportedWorkaroundForIssue375
+            ? ',unsupportedWorkaroundForIssue375='
+                '${widget.unsupportedWorkaroundForIssue375}'
+            : '';
+    final userAgent = widget.userAgent?.isNotEmpty == true
+        ? ',userAgent=${widget.userAgent}'
+        : '';
+
     return '[WebView:url=${widget.url}'
         ',aspectRatio=${widget.aspectRatio.toStringAsFixed(2)}'
         "${widget.autoResize ? ',autoResize=${widget.autoResize}' : ''}"
-        "${widget.debuggingEnabled ? ',debuggingEnabled=${widget.debuggingEnabled}' : ''}"
+        '$debuggingEnabled'
         "${!widget.js ? ',js=${widget.js}' : ''}"
-        "${widget.mediaPlaybackAlwaysAllow ? ',mediaPlaybackAlwaysAllow=${widget.mediaPlaybackAlwaysAllow}' : ''}"
-        "${!widget.unsupportedWorkaroundForIssue37 ? ',unsupportedWorkaroundForIssue37=${widget.unsupportedWorkaroundForIssue37}' : ''}"
-        "${!widget.unsupportedWorkaroundForIssue375 ? ',unsupportedWorkaroundForIssue375=${widget.unsupportedWorkaroundForIssue375}' : ''}"
-        "${widget.userAgent?.isNotEmpty == true ? ',userAgent=${widget.userAgent}' : ''}"
+        '$mediaPlaybackAlwaysAllow'
+        '$unsupportedWorkaroundForIssue37'
+        '$unsupportedWorkaroundForIssue375'
+        '$userAgent'
         ']';
   }
 
